sinks,avro-ocf: create Avro Writer/open File only when needed
All the writing of a Avro OCF sink happens from one worker. However,
before this change all workers would open a file, with the same file
name even.

This changes two things:
 - We make it explicit that we route all messages to one worker
 - We lazily create the Avro Writer when we receive the first messages

The type of avro_writer (Option<Result<>>) is a bit awkward, but it
reflects that fact that we try creating the writer once, and that can
result in an error. Before, we would try creating the writer outside the
sink function, and then inside check whether it was succesful.

The previous code would silently "continue" with the sink, only logging
and error once, when creating the writer failed. Now, we log an error
every time we try and receive the created writer, which will make the
error more noticeable in the logs.